chore: change mint api path
