--- packages/samples/sudt/src/dto/transfer.dto.ts
@@ -1,12 +1,0 @@
-import { TransactionSkeletonObject } from '@ckb-lumos/helpers'
-
-export interface TransferRequest {
-  from: string[]
-  to: string
-  amount: bigint
-  typeId: string
-}
-
-export interface TransferResponse {
-  skeleton: TransactionSkeletonObject
-}
